+++ cranangularclient/src/app/menu/menu.component.html
@@ -4,6 +4,7 @@
       Cranium
       <!-- user info -->
       <app-user-info></app-user-info>
+      <app-icon [icon]="'notification'" (click)="subscribeToPushNotifications()"></app-icon>
     </span>
   
     <!-- burger button -->
