+++ cranangularclient/src/app/subscriptions/subscriptions.component.ts
@@ -1,8 +1,9 @@
 
 import { Component, OnInit, Inject, ViewChild } from '@angular/core';
-
 import { SwPush } from '@angular/service-worker';
+
 import { NotificationService } from '../notification.service';
+import { LanguageService } from '../language.service';
 import { ICranDataService } from '../icrandataservice';
 import { CRAN_SERVICE_TOKEN } from '../cran-data.servicetoken';
 import { SubscriptionShort } from '../model/subscriptionshort';
@@ -26,12 +27,10 @@ export class SubscriptionsComponent implements OnInit {
 
 
   constructor(private swPush: SwPush, private notificationService: NotificationService,
-    @Inject(CRAN_SERVICE_TOKEN) private cranDataService: ICranDataService) {
+    @Inject(CRAN_SERVICE_TOKEN) private cranDataService: ICranDataService,
+      private ls: LanguageService) {
     if (this.swPush.isEnabled) {
       this.notificationEnabled = true;
-      this.swPush.messages.subscribe(this.messageSub);
-      this.swPush.notificationClicks.subscribe(this.notificationClicksSub);
-      this.checkForSubscripton();
     }
     this.getUsers();
   }
@@ -44,40 +43,6 @@ export class SubscriptionsComponent implements OnInit {
     this.users = subscriptions.data;
   }
 
-  private async checkForSubscripton() {
-    const serviceWorker: ServiceWorkerRegistration = await window.navigator.serviceWorker.ready;
-    const subscription: PushSubscription = await serviceWorker.pushManager.getSubscription();
-    if (subscription != null ) {
-      this.subscriptionJSON = JSON.stringify(subscription);
-    }
-  }
-
-  public async subscribeToPushNotifications()  {
-    try {
-      const subscription: PushSubscription = await this.swPush.requestSubscription({
-        serverPublicKey: this.VAPID_PUBLIC_KEY
-      });
-      this.subscriptionJSON = JSON.stringify(subscription);
-      try {
-        this.notificationService.emitLoading();
-        this.cranDataService.addPushRegistration(subscription);
-        this.notificationService.emitDone();
-      } catch (error) {
-        this.notificationService.emitError(error);
-      }
-    } catch (error) {
-    }
-  }
-
-  public messageSub(input: any) {
-    console.log('message received');
-    console.log(JSON.stringify(input));
-  }
-
-  public notificationClicksSub(input: any) {
-    console.log('notificationClicksSub');
-    console.log(JSON.stringify(input));
-  }
 
   public async sendNotification() {
     try {
